+++ client/package.json
@@ -19,6 +19,8 @@
     "react-bootstrap": "^2.6.0",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
+    "react-router": "^6.8.0",
+    "react-router-dom": "^6.8.0",
     "react-scripts": "4.0.3",
     "sass": "^1.56.2",
     "typescript": "~4.1.5"
