+++ client/src/RouteComponent.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+
+import Home from "./pages/Home";
+import Login from "./pages/Login";
+import NotFound from "./pages/NotFound";
+
+const RouteComponent = () => {
+	return (
+		<Router>
+			<Routes>
+				<Route path="/" element={<Home />} />
+				<Route path="/login" element={<Login />} />
+				<Route path="*" element={<NotFound />} />
+			</Routes>
+		</Router>
+	);
+};
+
+export default RouteComponent;
