+++ client/src/pages/Login.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function LoginPage() {
+	return <div></div>;
+}
