+++ client/src/pages/NotFound.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function NotFound() {
+	return <div>NotFound</div>;
+}
